r.hydrodem: sync to r.watershed
git-svn-id: https://svn.osgeo.org/grass/grass-addons@67237 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ grass7/raster/r.hydrodem/load.c
@@ -92,7 +92,7 @@ int load_map(int ele_fd)
     Rast_close(ele_fd);
     G_free(ele_buf);
 
-    G_debug(1, "%d non-NULL cells", n_points);
+    G_debug(1, "%lld non-NULL cells", (long long int)n_points);
 
     return 1;
 }
